bot/origami-invest-bot-main/requirements.txt
aiogram==3.4.0
python-dotenv==1.0.1
requests==2.31.0